requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
dj-email-url==1.0.6
Django==3.2.15
django-cache-url==3.4.4
django-debug-toolbar==3.2.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
environs==9.3.2
marshmallow==3.19.0
packaging==23.0
phonenumbers==8.13.4
Pillow==9.4.0
python-dotenv==0.21.0
pytz==2022.7.1
sqlparse==0.4.3
